requirements.txt
argh
autopep8>=1.3
google.cloud.logging
numpy>=1.14.0
petname
protobuf
pygtp>=0.4
pylint
sgf==0.5
six
tqdm>=4.17

# Model Zoo Requirements
httplib2
google-cloud-storage
oauth2client==4.1
google-api-python-client
kubernetes
google-auth
